[WARP-SPEED] SIC-2189 fix signal 11 in case of a panic
(cherry picked from commit 0676fb649f005f863f786ce638667612ec495a01)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/varada/dispatcher/warmup/warmers/StorageWarmerService.java
@@ -108,7 +108,7 @@ public class StorageWarmerService
     public int warmupOpen(int txId)
     {
         warmupClose(txId);
-        txId = storageEngine.warmupOpen(connectorSync.getCatalogSequence());
+        txId = (int) storageEngine.warmupOpen(connectorSync.getCatalogSequence());
         if (txId == INVALID_TX_ID) {
             throw new RuntimeException("failed to allocate tx for write");
         }
