+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/varada/storage/read/StorageReader.java
@@ -202,7 +202,7 @@ public class StorageReader
 
         if (queryParams.getNumMatchElements() > 0) {
             try {
-                matchTxId = storageEngine.matchOpen(queryParams.getTotalNumRecords(),
+                matchTxId = (int) storageEngine.matchOpen(queryParams.getTotalNumRecords(),
                         storageCollectorArgs.fileCookie(),
                         collectOpenResult.collectTxId(),
                         storageCollectorArgs.collectStoreBuff(),
